Add user option for pandas or numpy data types
This update should allow users the choice between numpy or pandas data
types when returning data from a file. Memory objects are assumed to work
exclusively in Numpy arrays, and the Memory wrapper in main will convert
those to Series or DataFrames depending on the type of data. The data in
the readers though must be managed by the objects themselves.

This also brings updated testing to hopefully improve stability, catch bugs
quicker, and allow tests to be easier to read.