Update documentation for PyPWA
This is a big overhaul to PyPWA's documentation. PyPWA now has proper
documentation that is listed on ReadTheDocs instead of just inline
documentation and the short how to install documentation that was
originally found on the ReadTheDocs. This is just the first major
batch of documentation, and more will come with time as we collect
feedback.