Patch process library to reuse Pipes
The pipes were being closed the way they're meant to be, but since
we relied on creating them and tearing them down inside loops, even
with them being closed the interpreter would eventually run out of
the number of file handles that it could track, overflowing the
interpreter and eventually causing the process to crash.

This solution bypasses that entire mess by creating a singleton
that stores each pair of faux pipes that have been created. When a pair
is unused, it'll return that pair, but if there are currently no unused
pairs, it'll return a new pair. This should mean that the number of
created pipes should stay low and manageable even if someone goes crazy
with creating pipes.